Test implementation that will write to a queue file, then send that file in a separate process.
This is broken - I've been on the fence with how to use the
multiprocessing module, etc. and in the end I think I'd be better off
not trying to control this from within the current relay.

I'm going to go further with this and have a new class that just
queues up files of data to send, and then a queue-runner cleans up
that queue for each destination.