This is the version that we're using in production.
It's a slightly older rev, but it turns out there's a bug in the use
of an open file list that needs to be resolved.

This is the same as the staging relay, with the difference being
that paused is always false, so spooling will continue when a remote
goes down.  If a remote is down when the relay starts, though, that
will prevent spooling.  That is a todo item.